Refactor: separate AdminMessages component into its own file
